+++ lib/model/Facility.php
@@ -9,18 +9,21 @@ class sspmod_perun_model_Facility implements sspmod_perun_model_HasId
 {
 	private $id;
 	private $name;
+	private $description;
 	private $entityId;
 
 	/**
 	 * sspmod_perun_model_Vo constructor.
 	 * @param int $id
 	 * @param string $name
+	 * @param string $description
 	 * @param string $shortName
 	 */
-	public function __construct($id, $name, $entityId)
+	public function __construct($id, $name, $description, $entityId)
 	{
 		$this->id = $id;
 		$this->name = $name;
+		$this->description = $description;
 		$this->entityId = $entityId;
 	}
 
@@ -35,6 +38,14 @@ class sspmod_perun_model_Facility implements sspmod_perun_model_HasId
 	public function getName()
 	{
 		return $this->name;
+	}
+
+	/**
+	 * @return string
+	 */
+	public function getDescription()
+	{
+		return $this->description;
 	}
 
 	/**