+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
  ## [Unreleased]
  [Added]
  - Added badges to README
+ - Added new property to Facility model: description
  
  ## [v1.0.0]